Refactor/rename.
Drop `parallel_execution` It’s just the `Executor` from now on.
Shorten the middleware filenames. On Tox, run import-order on python3.5
