+++ tox.ini
@@ -14,6 +14,7 @@ deps = flake8
 commands = flake8
 
 [testenv:import-order]
+basepython=python3.5
 deps = import-order
 commands = import-order graphql
